+++ packages/components/src/components.d.ts
@@ -8,6 +8,7 @@ import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
 import { Events, Props } from "./types";
 import { BalMode } from "./utils/config";
 import { BalCarouselItemData } from "./components/bal-carousel/bal-carousel.type";
+import { Frameworks } from "./components/docs/bal-doc-stackblitz/stackblitz.util";
 import { FileUploadRejectedFile } from "./components/form/bal-file-upload/bal-file-upload.type";
 import { OverlayEventDetail } from "./components/notice/bal-modal/bal-modal.type";
 import { LevelInfo } from "./components/bal-navigation/utils/level.utils";
@@ -66,6 +67,7 @@ export namespace Components {
           * Mode defines how the styles are loaded. With `css` each component loads his own styles and with `sass` the component styles needs to be imported with the file `global.components.sass`.
          */
         "mode": BalMode;
+        "ready": boolean;
         "setFocus": (elements: HTMLElement[]) => Promise<void>;
     }
     interface BalBadge {
@@ -678,10 +680,16 @@ export namespace Components {
     interface BalDocStackblitz {
         "component": string;
         "component2": string;
+        "framework": Frameworks;
+        "fullscreen": boolean;
+        "label": string;
+        "logo": boolean;
         "modules": string;
         "name2": string;
+        "primary": boolean;
         "template": string;
         "template2": string;
+        "visible": boolean;
     }
     interface BalDocSupportColor {
         "color": string;
@@ -1326,7 +1334,11 @@ export namespace Components {
         /**
           * Sets justify-content of the items to start, center, end, or space-between. Default is start
          */
-        "contentAlignment": Props.BalListContentSpacing;
+        "contentAlignment": Props.BalListContentAlignment;
+        /**
+          * Sets space to content of the accordion body
+         */
+        "contentSpace": Props.BalListContentSpacing;
     }
     interface BalListItemAccordionHead {
         /**
@@ -3619,6 +3631,7 @@ declare namespace LocalJSX {
          */
         "mode"?: BalMode;
         "onBalAppLoad"?: (event: BalAppCustomEvent<boolean>) => void;
+        "ready"?: boolean;
     }
     interface BalBadge {
         /**
@@ -4271,10 +4284,16 @@ declare namespace LocalJSX {
     interface BalDocStackblitz {
         "component": string;
         "component2": string;
+        "framework": Frameworks;
+        "fullscreen"?: boolean;
+        "label": string;
+        "logo"?: boolean;
         "modules": string;
         "name2": string;
+        "primary"?: boolean;
         "template": string;
         "template2": string;
+        "visible"?: boolean;
     }
     interface BalDocSupportColor {
         "color"?: string;
@@ -4951,7 +4970,11 @@ declare namespace LocalJSX {
         /**
           * Sets justify-content of the items to start, center, end, or space-between. Default is start
          */
-        "contentAlignment"?: Props.BalListContentSpacing;
+        "contentAlignment"?: Props.BalListContentAlignment;
+        /**
+          * Sets space to content of the accordion body
+         */
+        "contentSpace"?: Props.BalListContentSpacing;
     }
     interface BalListItemAccordionHead {
         /**