+++ packages/components/src/components/docs/bal-doc-stackblitz/stackblitz.const.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_EDITOR_TITLE = 'Baloise Design System Example'
+export const DEFAULT_EDITOR_DESCRIPTION = ''
+export const DEFAULT_BALOISE_VERSION = 'latest'
+export const PLACEHOLDER_IMPORT = '/** PLACEHOLDER FOR DESIGN SYSTEM IMPORTS */'