+++ packages/css/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @baloise/design-system-css
 
+## 12.4.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @baloise/design-system-fonts@12.4.1
+  - @baloise/design-system-tokens@12.4.1
+
+## 12.4.0
+
+### Patch Changes
+
+- [#1058](https://github.com/baloise-incubator/design-system/pull/1058) [`282355d61`](https://github.com/baloise-incubator/design-system/commit/282355d61f9e07882fca65a02b0108fc9e712397) Thanks [@hirsch88](https://github.com/hirsch88)! - set focus primary border on invalid form controls
+
+- Updated dependencies []:
+  - @baloise/design-system-fonts@12.4.0
+  - @baloise/design-system-tokens@12.4.0
+
 ## 12.3.0
 
 ### Patch Changes