+++ packages/fonts/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @baloise/design-system-fonts
 
+## 12.4.1
+
+## 12.4.0
+
 ## 12.3.0
 
 ## 12.2.0