+++ packages/output-targets/react/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @baloise/design-system-output-target-react
 
+## 12.4.1
+
+## 12.4.0
+
 ## 12.3.0
 
 ### Patch Changes